Solution. Initially I used a conditional statement to match the array elements in set A (good) expressed a boolean. So if true += 1 to happiness and vice versa for Set B -= 1 to happiness.
This worked for the first 3 test cases but the remaining 5 passed arrays that were much larger than the sets.

I was calculating the wrong score when using this method. After some research and review of the discussion boards I learned that the
happiness was just a score of the differences between the sets that would match the array elements.

Used list comprehension to work this out.